# server/swagger.yaml
openapi: 3.0.0
info:
  title: FloNeo LCNC Platform API
  description: RESTful API for the FloNeo Low-Code No-Code Platform
  version: 1.0.0
  contact:
    name: FloNeo Team
    email: support@floneo.com
servers:
  - url: http://localhost:5000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    App:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [Draft, Published, Active]
        ownerId:
          type: integer
        templateId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        template:
          $ref: '#/components/schemas/Template'
          nullable: true
    Template:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        preview_image:
          type: string
        category:
          type: string
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /api/apps:
    get:
      summary: List user's apps
      description: Retrieve apps owned by the authenticated developer with pagination and filtering
      tags:
        - Apps
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [Draft, Published, Active]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Apps retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      apps:
                        type: array
                        items:
                          $ref: '#/components/schemas/App'
                      pagination:
                        type: object
    post:
      summary: Create new app
      description: Create a new app for the authenticated developer
      tags:
        - Apps
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: "My Dashboard App"
                description:
                  type: string
                  example: "A custom dashboard application"
                templateId:
                  type: integer
                  example: 2
                  description: "Optional template ID to create app from template"
            examples:
              custom_app:
                summary: "Custom App (no template)"
                value:
                  name: "Custom Business App"
                  description: "A custom application built from scratch"
              template_app:
                summary: "App from Template"
                value:
                  name: "My Dashboard"
                  description: "Dashboard created from template"
                  templateId: 2
      responses:
        '201':
          description: App created successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
  /api/apps/{id}/status:
    get:
      summary: Get app status and metrics
      description: Retrieve app status and production metrics for dashboard
      tags:
        - Apps
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: App status retrieved successfully
        '404':
          description: App not found
        '401':
          description: Unauthorized
    put:
      summary: Update app status
      description: Update app status (Draft → Published → Active)
      tags:
        - Apps
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [Draft, Published, Active]
      responses:
        '200':
          description: App status updated successfully
        '400':
          description: Invalid status value
        '404':
          description: App not found
        '401':
          description: Unauthorized
...